Added buttons for controlling a song counter.
The Start Counter button starts (single click) or stops (double click)
a counter which tracks the song number within a tanda. The variable
that holds the song number is $tandaCounter. In the current
implementation, the counter keeps track of when a new tanda starts by
watching for tracks witha  "grouping" tag set to "Cortina".

To the left of the Start Counter button is a button which, when clicked,
toggles between 3-song tandas and 4-song tandas for tracks with a genre
of Tango.

To the right of Start Counter is a button which displays the song
number. If the counter gets out of synch with the currently playing
tanda, double-clicking it sets the counter to 0 while single-clicking it
can be used to add one to the counter until the correct number is shown.