changes to song counter to improve its performance even more in automix mode.
These changes seem to make the counter robust in automix mode. It has
been tested with skip silence in automix mode with the fade length set
to -20 s, a value no dj in their right mind would ever use. The counter
now updates fast in automix mode and is only slightly slower to update
when automix is turned off. The modifications include the following:

1. Removed -1 from "repeat_start_instant 1000ms -1". It doesn't seem
to be necessary.

2. It looks like load_pulse_active doesn't always play well with
automix mode, especially when fade length is set to a large negative
value. So, when automix mode is active, use a change in the value of
get_automix_position to detect that a new track has started playing. When
automix mode is off and decks are being loaded with tracks manually,
use load_pulse_active instead to detect track changes. This seems to
make the song counter more robust in automix.

3. Removed all the waitTime repeats as it seems they're not necessary
any more. This has the happy side effect of causing quicker counter
updates.

4. To correct a problem with the technique for detecting track changes
mentioned in 2, it was necessary to add a check to set the song counter
to 1 when the counter
is turned on and a non-cortina track is already playing. Thus, when you
start playing the first song of a tanda, it doesn't matter if you start
the song counter before or after the song starts playing. The original
implementation of the counter didn't have this problem since
load_pulse_active would fire to increment the counter from 0 to 1.

5. As this implementation doesn't seem to have an issue with double
counting, the double count check has been removed.

6. It was necessary to add a check to prevent the counter from
ticking up when automix mode is switched on and off.

7. The load_pulse_active time was reduced from 1500ms to 250ms, the
same as the original song counter commit.

8. Set the song counter to 0 when it is stopped. Probably not
necessary, but might be useful for video skins. E.g., this allows a
video skin to take some action if either a cortina is playing or the
counter is turned off simply by checking $tandaCounter with code like
"not var '$tandaCounter' ? ...".

9. On starting up, if genre check is turned on and you start playing a
track whose genre does not contain any of the words "tango", "vals",
"milonga", "alternative" or "cortina", the tanda length does not get
set and is blank wherever it is displayed, e.g., the button shows "1/".
To prevent this from happening, the tanda length is now set to 3, but
only  if it is found to be unset. This does not guarantee a correct
tanda length, but it is only for dealing with the first time it happens
and it is better to have the tanda length set to a positive value than
unset, which VDJScript interprets as 0.